project/src/requirements.txt
bcrypt
python-dotenv
mysql-connector-python